feat & fix: agregue y corregi una manera de agregar comandos de manera simple y rapida
