+++ src/components/CommandsContainer/commands/gitInit/gitInit.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function GitInit() {
+  return (
+    <div>
+      <p>git init</p>
+    </div>
+  );
+}
