+++ src/components/MenuMobile/MenuMobile.scss
@@ -1,5 +1,5 @@
 .MenuMobile {
   @include layout(0 20px);
 
-  padding-top: 4.2rem;
+  padding-top: 2rem;
 }
